chore: move frontend files to /client folder
