+++ client/components.json
