+++ client/eslint.config.mjs
