+++ client/next.config.ts
