+++ client/public/fonts/Satoshi/OTF/Satoshi-Medium.otf
