+++ client/public/fonts/Satoshi/OTF/Satoshi-MediumItalic.otf
