+++ client/src/hooks/use-mobile.ts
