+++ client/src/hooks/use-mounted.ts
