+++ client/src/i18n/config.ts
@@ -1,4 +1,4 @@
 export type Locale = (typeof locales)[number];
 
-export const locales = ['en', 'de'] as const;
+export const locales = ['en', 'pt-BR'] as const;
 export const defaultLocale: Locale = 'en';
