+++ client/src/messages/pt-BR/settings.json
